sending length/rows

+++ include/words.h
@@ -45,7 +45,7 @@ public:
     void incrementstateattemptandsuccess();
     static bool indictionary(std::string s);
     static WordValidation checkword(uint32_t* result,const std::string& guess,const std::string& correct);
-    static uint8_t getrowsforlength(uint64_t length);
+    static uint8_t getrowsforlength(size_t length);
     static bool isvalidlength(const std::string& str);
 private:
 protected:
